requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
PyJWT==2.8.0
requests==2.31.0
qiskit==0.45.0
qiskit-aer==0.13.0
qiskit-ibm-runtime==0.15.0
cryptography==41.0.7
numpy==1.24.3
